+++ temba/channels/types/plivo/tests.py
@@ -165,6 +165,30 @@ class PlivoTypeTest(TembaTest):
                 )
 
         with patch("requests.delete") as mock_delete:
-            response_body = MockResponse(200, json.dumps({"status": "ok"}))
             channel.get_type().deactivate(channel)
             mock_delete.assert_called_once()
+
+    def test_search(self):
+        self.login(self.admin)
+
+        search_url = reverse("channels.types.plivo.search")
+
+        with patch("requests.get") as mock_get:
+            mock_get.return_value = MockResponse(200, json.dumps({"objects": []}))
+
+            response = self.client.post(search_url, {"country": "US", "pattern": ""})
+
+            self.assertEqual(response.status_code, 200)
+            self.assertNotContains(response, "error")
+
+            # missing key to throw exception
+            mock_get.return_value = MockResponse(200, json.dumps({}))
+            response = self.client.post(search_url, {"country": "US", "pattern": ""})
+
+            self.assertEqual(response.status_code, 200)
+            self.assertContains(response, "error")
+
+            mock_get.return_value = MockResponse(400, "Bad request")
+            response = self.client.post(search_url, {"country": "US", "pattern": ""})
+
+            self.assertContains(response, "Bad request")
