+++ temba/channels/types/twilio_whatsapp/tests.py
@@ -63,7 +63,7 @@ class TwilioWhatsappTypeTest(TembaTest):
             self.assertTrue(response.context["account_trial"])
 
         with patch("temba.tests.twilio.MockTwilioClient.MockPhoneNumbers.list") as mock_search:
-            search_url = reverse("channels.channel_search_numbers")
+            search_url = reverse("channels.types.twilio.search")
 
             # try making empty request
             response = self.client.post(search_url, {})
@@ -71,26 +71,26 @@ class TwilioWhatsappTypeTest(TembaTest):
 
             # try searching for US number
             mock_search.return_value = [MockTwilioClient.MockPhoneNumber("+12062345678")]
-            response = self.client.post(search_url, {"country": "US", "area_code": "206"})
+            response = self.client.post(search_url, {"country": "US", "pattern": "206"})
             self.assertEqual(response.json(), ["+1 206-234-5678", "+1 206-234-5678", "+1 206-234-5678"])
 
             # try searching without area code
-            response = self.client.post(search_url, {"country": "US", "area_code": ""})
+            response = self.client.post(search_url, {"country": "US", "pattern": ""})
             self.assertEqual(response.json(), ["+1 206-234-5678", "+1 206-234-5678", "+1 206-234-5678"])
 
             mock_search.return_value = []
-            response = self.client.post(search_url, {"country": "US", "area_code": ""})
+            response = self.client.post(search_url, {"country": "US", "pattern": ""})
             self.assertEqual(
                 response.json()["error"], "Sorry, no numbers found, please enter another area code and try again."
             )
 
             # try searching for non-US number
             mock_search.return_value = [MockTwilioClient.MockPhoneNumber("+442812345678")]
-            response = self.client.post(search_url, {"country": "GB", "area_code": "028"})
+            response = self.client.post(search_url, {"country": "GB", "pattern": "028"})
             self.assertEqual(response.json(), ["+44 28 1234 5678", "+44 28 1234 5678", "+44 28 1234 5678"])
 
             mock_search.return_value = []
-            response = self.client.post(search_url, {"country": "GB", "area_code": ""})
+            response = self.client.post(search_url, {"country": "GB", "pattern": ""})
             self.assertEqual(
                 response.json()["error"], "Sorry, no numbers found, please enter another pattern and try again."
             )
