+++ temba/settings_common.py
@@ -449,8 +449,6 @@ PERMISSIONS = {
         "create_caller",
         "errors",
         "facebook_whitelist",
-        "search_vonage",
-        "search_numbers",
     ),
     "channels.channellog": ("connection",),
     "channels.channelevent": ("api", "calls"),
@@ -665,8 +663,6 @@ GROUP_PERMISSIONS = {
         "channels.channel_delete",
         "channels.channel_list",
         "channels.channel_read",
-        "channels.channel_search_vonage",
-        "channels.channel_search_numbers",
         "channels.channel_update",
         "channels.channelevent.*",
         "channels.channellog_list",
@@ -781,7 +777,6 @@ GROUP_PERMISSIONS = {
         "channels.channel_delete",
         "channels.channel_list",
         "channels.channel_read",
-        "channels.channel_search_numbers",
         "channels.channel_update",
         "channels.channelevent.*",
         "flows.flow.*",
