+++ templates/channels/types/plivo/claim.haml
@@ -1,10 +1,10 @@
 -extends "channels/channel_claim_number.html"
-
--load compress temba
 -load i18n
-
 
 -block country-select
   %temba-select#country(name="country" label='{{_("Country")|escapejs}}')
     -for country in search_countries
       %temba-option(name="{{country.label}}" value="{{country.key}}")
+
+-block search-pattern
+  %temba-textinput#pattern(type="text" maxlength="7" name="pattern" label='{{_("Pattern")|escapejs}}')
